app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
google-api-python-client==2.147.0
google-auth==2.35.0
